Change flash_program API to take a *u8
Back when the LPC55 ROM API were being merged there was some discussion
on the proper types for the programming APIs (c.f.
https://github.com/oxidecomputer/hubris/pull/132). Specifically, there
were concerns about whether the APIs should take a `*u8` or a `*u32`
since there was a mismatch in a few places in the documentation.
`*u32` was the more conservative option so that was used.

Hubris mostly works on `u8` arrays. These need to be converted to `u32`
arrays to work with the programming APIs. This is a pain. The underlying
ROM does support `u8` (i.e. unaligned) arrays so just change the API
to what we're actually using.

+++ drv/lpc55-romapi/src/lib.rs
@@ -131,7 +131,7 @@ struct Version1DriverInterface {
     flash_program: unsafe extern "C" fn(
         config: &mut FlashConfig,
         start: u32,
-        src: *mut u32,
+        src: *mut u8,
         length: u32,
     ) -> u32,
     /// flash_verify_erase: Verify that the region is actually erased
@@ -528,7 +528,7 @@ pub unsafe fn flash_erase(addr: u32, len: u32) -> Result<(), FlashStatus> {
 
 pub unsafe fn flash_write(
     addr: u32,
-    buffer: *mut u32,
+    buffer: *mut u8,
     len: u32,
 ) -> Result<(), FlashStatus> {
     //   XXX More validation of buffer?
